EPMLABSBRN-306 move interfaces to services
