+++ src/main/kotlin/com/epam/brn/csv/converter/DataLoadingBeanProvider.kt
@@ -1,0 +1,10 @@
+package com.epam.brn.csv.converter
+
+import org.springframework.data.repository.CrudRepository
+
+interface DataLoadingBeanProvider<Csv, Entity> {
+    fun shouldProcess(fileName: String): Boolean
+    fun repository(): CrudRepository<Entity, Long>
+    fun objectReaderProvider(): ObjectReaderProvider<Csv>
+    fun converter(): CsvToEntityConverter<Csv, Entity>
+}
