+++ src/main/kotlin/com/epam/brn/csv/converter/ObjectReaderProvider.kt
@@ -1,0 +1,8 @@
+package com.epam.brn.csv.converter
+
+import com.fasterxml.jackson.databind.ObjectReader
+
+@FunctionalInterface
+interface ObjectReaderProvider<Type> {
+    fun <Type> objectReader(): ObjectReader
+}
