--- src/main/kotlin/com/epam/brn/csv/converter/Uploader.kt
@@ -1,9 +1,0 @@
-package com.epam.brn.csv.converter
-
-import com.fasterxml.jackson.databind.ObjectReader
-
-interface Uploader<Csv, Entity> : CsvToEntityConverter<Csv, Entity> {
-    fun shouldProcess(fileName: String): Boolean
-    fun save(entity: Entity)
-    fun objectReader(): ObjectReader
-}
